Abandon repeat files and improve project structure
