+++ demo/aes/aes_encrypt.go
@@ -1,4 +1,4 @@
-package util
+package aes
 
 import (
 	"bytes"
