+++ demo/aes/aes_encrypt_test.go
@@ -1,4 +1,4 @@
-package util
+package aes
 
 import (
 	"fmt"
