+++ demo/opts/main.go
@@ -1,4 +1,4 @@
-package main
+package opts
 
 import (
 	"flag"
@@ -13,7 +13,7 @@ var lang = flag.String("lang", "golang", "the lang of the program")
 var age = flag.Int64("age", 18, "the age of the user")
 var safemod = flag.Bool("safemod", true, "whether using safemode")
 
-func main() {
+func getOpts() {
 	//flag.Parse()
 	fmt.Println("lang is: ", *lang)
 	fmt.Println("age is: ", *age)
