--- demo/pgops4/main.go
@@ -1,69 +1,0 @@
-package main
-
-// schema we can use along with some select statements
-// create table test ( gopher_id int, created timestamp );
-// select * from test order by created asc limit 1;
-// select * from test order by created desc limit 1;
-// select count(created) from test;
-
-import (
-	"database/sql"
-	"fmt"
-	"log"
-	"time"
-
-	"github.com/lib/pq"
-)
-
-const (
-	gophers = 10
-	entries = 10000
-)
-
-// Dump table strcuture
-// pg_dump -U postgres -O test  -t test>test.sql
-// pg copy load dara
-func main() {
-	// create string to pass
-	// About 0.15s
-	fmt.Printf("StartTime: %v\n", time.Now())
-	pgcopy()
-	fmt.Printf("StopTime: %v\n", time.Now())
-}
-
-func pgcopy() {
-	// lazily open db (doesn't truly open until first request)
-	db, err := sql.Open("postgres", "host=localhost user=postgres dbname=test sslmode=disable")
-	if err != nil {
-		log.Fatal(err)
-	}
-	txn, err := db.Begin()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	stmt, err := txn.Prepare(pq.CopyIn("test", "gopher_id", "created"))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	for i := 1; i <= gophers*entries; i++ {
-		_, err = stmt.Exec(i, time.Now().Format("2006-01-02 15:04:05.999999999Z07:00"))
-		if err != nil {
-			log.Fatal(err)
-		}
-	}
-
-	err = stmt.Close()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = txn.Commit()
-	if err != nil {
-		log.Fatal(err)
-	}
-	// close db
-	db.Close()
-
-}
