+++ demo/sql-transformer/util_test.go
@@ -1,4 +1,4 @@
-package gsql
+package transformer
 
 import (
 	"fmt"
